requirements.txt
mysql-connector-python~=8.2.0
flask~=3.0.0